requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.1
Django==3.2.10
django-allauth==0.47.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-dynamic-fields==0.3.1
drf-nested-routers==0.93.4
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.1.1
packaging==21.3
Pillow==9.1.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.9
python-decouple==3.6
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejwt==2.0.1
sqlparse==0.4.2
typing==3.7.4.3
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
